refactor(pwa): extract client registry and eliminate composerBridge hack
