+++ packages/pwa/src/stores/index.ts
@@ -10,7 +10,7 @@ import type { PlatformPorts, ChatComposer } from '@clawwork/core';
 import type { ModelCatalogEntry, AgentInfo, DebugEvent } from '@clawwork/shared';
 import { ports } from '../platform/index.js';
 import { getIdentity } from '../persistence/db.js';
-import { getClient } from '../gateway/client.js';
+import { getClient } from '../gateway/client-registry.js';
 import { reportDebugEvent } from '../lib/debug.js';
 import i18next from 'i18next';
 import { toast } from 'sonner';
@@ -112,10 +112,6 @@ const sessionSync = createSessionSync({
   }),
 });
 
-export const composerBridge = {
-  markAbortedByUser: (_taskId: string): void => {},
-};
-
 let _dispatcher: ReturnType<typeof createGatewayDispatcher> | null = null;
 let hydrationReady: Promise<void> | null = null;
 
@@ -175,7 +171,7 @@ function getComposer(): ChatComposer {
       getTaskStore: () => taskStoreApi.getState(),
       getMessageStore: () => messageStoreApi.getState(),
       persistMessage: (...args) => getPorts().persistence.persistMessage(...args),
-      markAbortedByUser: (taskId) => composerBridge.markAbortedByUser(taskId),
+      markAbortedByUser: (taskId) => getDispatcher().markAbortedByUser(taskId),
       compactSession: (gwId, sk) => {
         const c = getClient(gwId);
         if (!c) return Promise.resolve({ ok: false, error: 'Gateway not found' });
